-Multithreading support:   -Audio plays on separate thread (work in progress)   -Files load in a thread pool, so a simple loading screen was possible   -Graphic processes are will be transferred to a separate thread next time -More guns -Everything scales now (game works on any resolution) -Settings actually change stuff now -Game prints a debug log and save data to /My Games/WarmVector

+++ src/Util/Point.java
@@ -8,9 +8,9 @@ public class Point {
 
     // Simple custom point class
 
-    public float x, y;
+    public double x, y;
 
-    public Point(float x, float y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
